+++ main.go
@@ -8,17 +8,25 @@ import (
 	"math/rand/v2"
 	"os"
 
+	system_split "github.com/kurogee/wagyu-script/system-split"
+
 	array_pack "github.com/kurogee/wagyu-script/array"
 	date_pack "github.com/kurogee/wagyu-script/date"
 	file_pack "github.com/kurogee/wagyu-script/file"
 	string_pack "github.com/kurogee/wagyu-script/string"
 	get_pack "github.com/kurogee/wagyu-script/get"
 	regex_pack "github.com/kurogee/wagyu-script/regex"
+	dict_pack "github.com/kurogee/wagyu-script/dict"
+	// http_pack "github.com/kurogee/wagyu-script/http"
 
 	math_sharp_functions "github.com/kurogee/wagyu-script/maths"
 
 	"github.com/Knetic/govaluate"
 )
+
+var split = system_split.Split
+var divide_split = system_split.Divide_split
+var take_off_quotation = system_split.Take_off_quotation
 
 type Parse struct {
 	parsed []string
@@ -44,6 +52,8 @@ var packages = Package{
 	"array": array_pack.Run,
 	"string": string_pack.Run,
 	"regex": regex_pack.Run,
+	"dict": dict_pack.Run,
+	// "http": http_pack.Run,
 }
 
 var packages_with_functions = Package_with_functions{
@@ -57,15 +67,8 @@ var packages_sharp_functions = Package_sharp_functions{
 	"string": string_pack.Sharp,
 	"array": array_pack.Sharp,
 	"file": file_pack.Sharp,
+	"dict": dict_pack.Sharp,
 }
-
-var replace_chars map[string]string = map[string]string{
-	"\\;" : "__SEMICOLON__",
-	"\\\"" : "__DOUBLE_QUOTATION__",
-	"\\'" : "__SINGLE_QUOTATION__",
-	"\\`" : "__BACK_QUOTATION__",
-	"\\\\\\" : "__BACKSLASH__",
-};
 
 func contains(s []string, e string) bool {
 	for _, a := range(s) {
@@ -75,162 +78,6 @@ func contains(s []string, e string) bool {
 	}
 
 	return false
-}
-
-func split(input string) []map[string]bool {
-	var tokens []string
-	var buffer strings.Builder
-	var quoteChar rune
-	var parenStack []rune
-	
-	// 違う種類の括弧が重なっているところがあったら、それを離す
-	// 例: ({ → ( { にする など
-	re := regexp.MustCompile(`([\(\{\[\)\}\]])\s+([\(\{\[\)\}\]])`)
-	input = re.ReplaceAllString(input, "$1 $2")
-
-	// \\～を一時的にreplace_charsに置き換える
-	for key, value := range(replace_chars) {
-		input = strings.ReplaceAll(input, key, value)
-	}
-
-	inQuote := false
-	inQuotes := []bool{}
-
-	inParen := false
-	inFunctionCall := false
-
-	// Regular expressions for matching parentheses and braces
-	parenRegex := regexp.MustCompile(`^[\(\)\{\}]$`)
-	functionCallRegex := regexp.MustCompile(`^#\w*\(`)
-
-	for _, char := range input {
-		switch {
-		case inQuote:
-			buffer.WriteRune(char)
-			if char == quoteChar {
-				inQuote = false
-
-				inQuotes = append(inQuotes, true)
-
-				tokens = append(tokens, buffer.String())
-				buffer.Reset()
-			}
-
-		case inParen:
-			buffer.WriteRune(char)
-			if char == parenStack[len(parenStack)-1] && len(parenStack) > 0 {
-				parenStack = parenStack[:len(parenStack)-1]
-				if len(parenStack) == 0 {
-					inParen = false
-					if inFunctionCall {
-						tokens = append(tokens, buffer.String())
-						buffer.Reset()
-						inFunctionCall = false
-
-						inQuotes = append(inQuotes, false)
-					}
-				}
-			} else if parenRegex.MatchString(string(char)) {
-				parenStack = append(parenStack, matchingParen(char))
-			}
-
-		case inFunctionCall:
-			buffer.WriteRune(char)
-			if char == '(' {
-				parenStack = append(parenStack, ')')
-				inParen = true
-			} else if char == ')' && len(parenStack) == 0 {
-				inFunctionCall = false
-
-				tokens = append(tokens, buffer.String())
-				buffer.Reset()
-
-				inQuotes = append(inQuotes, false)
-			}
-
-		default:
-			switch char {
-			case ' ', '\t', '\n':
-				if buffer.Len() > 0 {
-					tokens = append(tokens, buffer.String())
-					buffer.Reset()
-
-					inQuotes = append(inQuotes, false)
-				}
-
-			case '\'', '"', '`':
-				inQuote = true
-				quoteChar = char
-				buffer.WriteRune(char)
-
-			case '(', '{':
-				inParen = true
-				parenStack = append(parenStack, matchingParen(char))
-				buffer.WriteRune(char)
-
-			default:
-				buffer.WriteRune(char)
-				if functionCallRegex.MatchString(buffer.String()) {
-					inFunctionCall = true
-				}
-			}
-		}
-	}
-
-	if buffer.Len() > 0 {
-		tokens = append(tokens, buffer.String())
-
-		if inQuote {
-			inQuotes = append(inQuotes, true)
-		} else {
-			inQuotes = append(inQuotes, false)
-		}
-	}
-	
-	// tokensのエスケープシーケンスを元に戻す
-	for key, value := range(replace_chars) {
-		for i, token := range(tokens) {
-			tokens[i] = strings.ReplaceAll(token, value, key)
-		}
-	}
-
-	// tokensをmapに変換し、inQuotesと合わせる
-	var tokens_map []map[string]bool
-	for i, token := range(tokens) {
-		tokens_map = append(tokens_map, map[string]bool{take_off_quotation(token): inQuotes[i]})
-	}
-
-	return tokens_map
-}
-
-// splitの返り値から、文字列とboolを別々にする関数
-func divide_split(tokens []map[string]bool) ([]string, []bool) {
-	var divided_tokens []string
-	var divided_tokens_bool []bool
-	for _, token := range(tokens) {
-		for key := range(token) {
-			divided_tokens = append(divided_tokens, key)
-		}
-		for _, val := range(token) {
-			divided_tokens_bool = append(divided_tokens_bool, val)
-		}
-	}
-
-	return divided_tokens, divided_tokens_bool
-}
-
-func matchingParen(char rune) rune {
-	switch char {
-	case '(':
-		return ')'
-	case '{':
-		return '}'
-	case ')':
-		return '('
-	case '}':
-		return '{'
-	}
-	return char
 }
 
 func giveSymbols(input string) string {
@@ -255,6 +102,14 @@ func parser(code string) (p Parse) {
 	code = re.ReplaceAllString(code, "")
 
 	var mem string = strings.ReplaceAll(code, " ", "")
+
+	// #関数名(引数)で、#関数名 (引数) と書かれていたら、間のスペースを削除
+	re = regexp.MustCompile(`(#[^\(]*) (\()`)
+	code = re.ReplaceAllString(code, "$1$2")
+
+	// (sharp|fnc) 関数名(引数) と書かれていたら、間にスペースを入れる
+	re = regexp.MustCompile(`(sharp|fnc) ([^\(]+)\(`)
+	code = re.ReplaceAllString(code, "$1 $2 (")
 	
 	if mem == "" {
 		p.parsed = []string{"None", ""}
@@ -343,22 +198,6 @@ func variables_replacer(variables *map[string]string, target string, target_in_q
 	// 
 	// 	return target
 	// }
-
-	return target
-}
-
-func take_off_quotation(target string) string {
-	if strings.HasPrefix(target, "'") && strings.HasSuffix(target, "'") {
-		// 一番外側のシングルクォーテーションを取り除く
-		re := regexp.MustCompile(`^'|'$`)
-		return re.ReplaceAllString(target, "")
-	} else if strings.HasPrefix(target, "\"") && strings.HasSuffix(target, "\"") {
-		re := regexp.MustCompile(`^"|"$`)
-		return re.ReplaceAllString(target, "")
-	} else if strings.HasPrefix(target, "`") && strings.HasSuffix(target, "`") {
-		re := regexp.MustCompile(`^`+"`"+`|`+"`"+`$`)
-		return re.ReplaceAllString(target, "")
-	}
 
 	return target
 }